+++ validator.js
@@ -402,7 +402,7 @@ function isURL(url, options) {
     }
   }
 
-  if (!isIP(host) && !isFDQN(host, options) && (!ipv6 || !isIP(ipv6, 6)) && host !== 'localhost') {
+  if (!isIP(host) && !isFDQN(host, options) && (!ipv6 || !isIP(ipv6, 6))) {
     return false;
   }