+++ src/lib/isURL.js
@@ -61,6 +61,10 @@ export default function isURL(url, options) {
   }
   url = split.join('://');
 
+  if (url === '') {
+    return false;
+  }
+
   split = url.split('/');
   url = split.shift();