+++ validator.js
@@ -1233,7 +1233,7 @@ function normalizeEmail(email, options) {
   return parts.join('@');
 }
 
-var version = '7.2.0';
+var version = '8.0.0';
 
 var validator = {
   version: version,