+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - `isURL()` now requires the `require_tld: false` option to validate `localhost`
   ([#675](https://github.com/chriso/validator.js/issues/675))
+- `isURL()` now rejects URLs that are protocol only
+  ([#642](https://github.com/chriso/validator.js/issues/642))
 
 #### 7.2.0